refactor: move reusable components
