+++ lact-gui/src/app/components/info_row.rs
