+++ lact-gui/src/app/components/page_section.rs
