+++ lact-gui/src/app/pages/oc_page/performance_frame.rs
@@ -3,7 +3,7 @@ mod heuristics_list;
 use super::OcPageMsg;
 use crate::{
     APP_BROKER, I18N,
-    app::{msg::AppMsg, page_section::PageSection},
+    app::{components::page_section::PageSection, msg::AppMsg},
 };
 use amdgpu_sysfs::gpu_handle::{PerformanceLevel, power_profile_mode::PowerProfileModesTable};
 use gtk::{
